requirements.txt
pyyaml>=6.0
python-magic-bin>=0.4.14
colorama>=0.4.6
